testovací raspberry/___sensation/requirements.txt
appdirs>=1.0
argcomplete>=1.0
pyserial>=3.0
bcf